--- src/js/popup/fuck.js
@@ -1,5 +1,0 @@
-const convert = () => {
-  console.log('sex')
-}
-
-module.exports = convert
