--- src/js/popup/sex.js
@@ -1,5 +1,0 @@
-const convert = () => {
-  console.log('sex')
-}
-
-module.exports = convert
