+++ src/js/popup/test.js
@@ -1,0 +1,3 @@
+module.exports = () => {
+  console.log('is called ?')
+}
