+++ src/sidenav.html
@@ -25,7 +25,7 @@
               </select>
             </div>
             <input
-              id="from-value"
+              id="fromText"
               class="input100"
               type="text"
               name="name"
